Split happy-core into happy-cli and happy-grammar
