+++ cabal.project
@@ -1,8 +1,9 @@
 packages:
- packages/frontend,
- packages/tabular,
  packages/backend,
  packages/backend-glr,
- packages/core,
- packages/test,
- packages/happy
+ packages/cli,
+ packages/frontend,
+ packages/grammar,
+ packages/happy,
+ packages/tabular,
+ packages/test
