+++ packages/cli/Setup.hs
