+++ packages/happy/happy.cabal
@@ -119,7 +119,8 @@ executable happy
                  containers >= 0.4.2,
                  mtl >= 2.2.1,
                  transformers >= 0.5,
-                 happy-core == 1.21.0,
+                 happy-grammar == 1.21.0,
+                 happy-cli == 1.21.0,
                  happy-frontend == 1.21.0,
                  happy-tabular == 1.21.0,
                  happy-backend == 1.21.0,
